Use theme variables across panels
