+++ css/style.css
@@ -3,21 +3,25 @@ body {
   margin: 0;
   padding: 0;
   font-family: Arial, sans-serif;
-  background-color: #121212;
-  color: #e0e0e0;
+  background-color: var(--bg-color);
+  color: var(--text-color);
 }
 
-/* Default panel colors */
+/* Default theme variables */
 :root {
-  --panel-bg: #1e1e2f;
-  --text-color: #e0e0e0;
+  --bg-color: #1a1a2e;
+  --panel-color: #292b3d;
+  --text-color: #f0f0f0;
+  --button-bg: #444;
+  --button-text: #ffffff;
+  --button-hover-bg: #666;
 }
 .category-panel {
   position: fixed;
   left: 0;
   top: 0;
   width: 220px;
-  background-color: #1e1e2f;
+  background-color: var(--panel-color);
   border-right: 2px solid #444;
   padding: 15px;
   height: 100vh;
@@ -30,11 +34,11 @@ body {
 #roleDefinitionsPanel {
   width: 460px;
   max-width: 100%;
-  background-color: #1a1a2e;
+  background-color: var(--panel-color);
   padding: 24px;
   overflow-y: auto;
   scroll-behavior: smooth;
-  color: #d6d6d6;
+  color: var(--text-color);
   font-family: 'Segoe UI', sans-serif;
   font-size: 16px;
   line-height: 1.6;
@@ -77,7 +81,7 @@ body.light-mode #closeRoleDefinitionsBtn {
   }
 
   .category-sidebar #categoryPanel {
-    background-color: var(--panel-bg, #1e1e1e);
+    background-color: var(--panel-color, #1e1e1e);
     max-width: 450px;
     width: 90%;
     max-height: 90vh;
@@ -195,40 +199,58 @@ body.theme-rainbow #mainCategoryList button.active {
 body.dark-mode {
   background-color: #121212;
   color: #e0e0e0;
-  --panel-bg: #1e1e2f;
+  --panel-color: #1e1e2f;
   --text-color: #e0e0e0;
+  --button-bg: #7c5fe9;
+  --button-text: #ffffff;
+  --button-hover-bg: #9b84ff;
 }
 body.light-mode {
   background-color: #b8c5b8;
   color: #2f4f2f;
-  --panel-bg: #a9b8a9;
+  --panel-color: #a9b8a9;
   --text-color: #2f4f2f;
+  --button-bg: #558c75;
+  --button-text: #ffffff;
+  --button-hover-bg: #446d60;
 }
 body.theme-blue {
   background-color: #001933;
   color: #fff;
-  --panel-bg: #002244;
+  --panel-color: #002244;
   --text-color: #fff;
+  --button-bg: #0055aa;
+  --button-text: #ffffff;
+  --button-hover-bg: #003377;
 }
 body.theme-echoes-beyond {
   /* Outer Wilds inspired colors */
   background-color: #14213d;
   color: #fca311;
-  --panel-bg: #1b263b;
+  --panel-color: #1b263b;
   --text-color: #fca311;
+  --button-bg: #334863;
+  --button-text: #fca311;
+  --button-hover-bg: #1f3352;
 }
 body.theme-love-notes-lipstick {
   /* Monster Prom inspired colors */
   background-color: #311847;
   color: #ff6bd6;
-  --panel-bg: #3f206b;
+  --panel-color: #3f206b;
   --text-color: #ff6bd6;
+  --button-bg: #ff6bd6;
+  --button-text: #311847;
+  --button-hover-bg: #c44fb0;
 }
 body.theme-rainbow {
   background: linear-gradient(135deg, #ff9a9e, #fad0c4, #fbc2eb, #a18cd1);
   color: #000;
-  --panel-bg: #fff;
+  --panel-color: #fff;
   --text-color: #000;
+  --button-bg: rgba(255, 255, 255, 0.9);
+  --button-text: #000;
+  --button-hover-bg: rgba(255, 255, 255, 0.7);
 }
 
 /* Theme-specific panel colors */
@@ -369,8 +391,8 @@ button {
   font-weight: bold;
   border: none;
   border-radius: 5px;
-  background-color: #444;
-  color: white;
+  background-color: var(--button-bg);
+  color: var(--button-text);
   transition: background-color 0.3s;
 }
 
@@ -383,6 +405,8 @@ textarea {
   padding: 8px;
   border-radius: 4px;
   border: 1px solid #888;
+  background-color: var(--panel-color);
+  color: var(--text-color);
   font-family: inherit;
 }
 
@@ -522,17 +546,17 @@ body.theme-rainbow #comparisonResult {
 
 
 .password-modal {
-  background-color: #1e1e2f;
+  background-color: var(--panel-color);
   padding: 20px;
   border-radius: 8px;
-  color: #fff;
+  color: var(--text-color);
 }
 
 #surveyIntro .intro-modal {
-  background-color: #1e1e2f;
+  background-color: var(--panel-color);
   padding: 20px;
   border-radius: 8px;
-  color: #fff;
+  color: var(--text-color);
   display: flex;
   flex-direction: column;
   gap: 10px;
@@ -557,7 +581,7 @@ body.theme-rainbow #comparisonResult {
   left: 0;
   height: 100vh;
   width: 320px;
-  background-color: var(--panel-bg, #1e1e1e);
+  background-color: var(--panel-color, #1e1e1e);
   color: var(--text-color, #ffffff);
   overflow-y: auto;
   padding: 24px 20px;
@@ -568,7 +592,7 @@ body.theme-rainbow #comparisonResult {
 }
 
 #categoryPanel {
-  background-color: var(--panel-bg, #fff);
+  background-color: var(--panel-color, #fff);
   padding: 24px;
   border-radius: 8px;
   color: var(--text-color, #2c442e);
@@ -773,7 +797,7 @@ button {
   overflow: hidden;
 }
 button:hover {
-  background-color: #666;
+  background-color: var(--button-hover-bg);
   box-shadow: 0 0 8px rgba(255, 255, 255, 0.6);
 }
 
@@ -1008,6 +1032,8 @@ body.light-mode .static-rating-legend {
   max-width: 600px;
   box-sizing: border-box;
   margin: 0 auto;
+  background-color: var(--panel-color);
+  color: var(--text-color);
 }
 
 #kinkList select {
