+++ js/theme.js
@@ -1,0 +1,13 @@
+export function initTheme() {
+  const themeSelector = document.getElementById('themeSelector');
+  const savedTheme = localStorage.getItem('selectedTheme') || 'dark-mode';
+  document.body.className = savedTheme;
+  if (themeSelector) {
+    themeSelector.value = savedTheme;
+    themeSelector.addEventListener('change', () => {
+      const selectedTheme = themeSelector.value;
+      document.body.className = selectedTheme;
+      localStorage.setItem('selectedTheme', selectedTheme);
+    });
+  }
+}
