+++ kink-list.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Your Kink List</title>
   <link rel="stylesheet" href="css/style.css" />
+  <link rel="stylesheet" href="css/theme.css" />
 </head>
 <body class="dark-mode">
   <div class="scroll-container scrollable-panel">
@@ -23,6 +24,8 @@
 
     <script src="js/template-survey.js"></script>
     <script type="module">
+      import { initTheme } from './js/theme.js';
+      initTheme();
 
       const RATING_LABELS = {
         0: 'Hard No',
