+++ your-roles.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Your Role Matches</title>
   <link rel="stylesheet" href="css/style.css" />
+  <link rel="stylesheet" href="css/theme.css" />
 </head>
 <body class="dark-mode">
   <div class="scroll-container scrollable-panel">
@@ -21,6 +22,8 @@
 
   <script type="module">
     import { calculateRoleScores } from './js/calculateRoleScores.js';
+    import { initTheme } from './js/theme.js';
+    initTheme();
 
     function flattenSurvey(survey) {
       const items = [];
